get prior CTS profile and covariance matrix from single-cell data

It calculates prior CTS profile and covariance matrix from single-cell data. The output can serve as hyper-parameters in bMIND. Only genes with positive definite covariance matrix are outputted.

get_prior(sc, meta_sc)

Arguments

sc

single-cell count matrix, gene x cell.

meta_sc

data.frame for meta of cells (cell x features, including columns `sample` (sample ID), `cell_type`).

Value

A list containing

profile

CTS profile matrix (gene x cell type), in log2(CPM + 1) scale.

covariance

CTS covariance matrix (gene x cell type x cell type).

References

Wang, Jiebiao, Kathryn Roeder, and Bernie Devlin. "Bayesian estimation of cell-type-specific gene expression per bulk sample with prior derived from single-cell data." bioRxiv (2020).